Move run internals into run.lib package
Refactor: relocate internal run modules into a new swanlab.sdk.internal.run.lib package (async_task, consumer, factory, record_builder, utils_fmt) and add lib/__init__.py. Update swanlab.sdk.internal.run.__init__ to import the moved components from .lib and adjust imports in tests and benchmark helper files to the new paths. Minor fix in factory module imports to reference run.config directly. This keeps run's internal helpers grouped under run.lib and updates all references accordingly.
