+++ tests/unit/sdk/internal/run/lib/test_async_task.py
@@ -12,7 +12,7 @@ import warnings
 
 import pytest
 
-from swanlab.sdk.internal.run.async_task import AsyncTaskManager
+from swanlab.sdk.internal.run.lib.async_task import AsyncTaskManager
 
 
 # spawn/fork 模式要求 func 可 pickle，lambda 和局部闭包不可 pickle，
