+++ tests/unit/sdk/internal/run/lib/test_run_fmt.py
@@ -7,7 +7,7 @@
 
 import pytest
 
-import swanlab.sdk.internal.run.utils_fmt as fmt
+import swanlab.sdk.internal.run.lib.utils_fmt as fmt
 
 # ─────────────────────────── flatten_dict ────────────────────────────
